Fix forward sarch
For backward search in skim use

in the preferances pane of Skim, navigate to Sync tab. There, in the PDF-TeX Sync support, make preset as custom, command as nvim(use vim if you use vim editor), and set arguments as --headless -c "VimtexInverseSearch %line %file".
This works also with neovide

Signed-off-by: Marcello Seri <marcello.seri@gmail.com>